requirements.txt
langextract
tabulate
python-dotenv
